Settings now displays a selector to select the role creation role
