+++ packages/trpc/src/utils.ts
@@ -1,5 +1,9 @@
 import { camelCase, isArray, isObject, transform } from 'lodash';
 
+import type { ApiGuild } from './routes/guilds';
+import type { ApiRole } from './routes/roles';
+import type { DiscordMember } from './routes/users';
+
 /* eslint-disable import/prefer-default-export */
 export const camelize = (obj: unknown) => {
   if (!isObject(obj)) return obj;
@@ -13,5 +17,34 @@ export const camelize = (obj: unknown) => {
         ? camelize(value as Record<string, unknown>)
         : value;
     },
+  );
+};
+
+export const highestRole = (roles: ApiRole[], member: DiscordMember) => {
+  const role = roles.reduce((prev, curr) => {
+    if (curr.position > prev.position && member.roles.includes(curr.id)) {
+      return curr;
+    }
+
+    return prev;
+  });
+
+  return role;
+};
+
+export const userIsAdmin = (
+  roles: ApiRole[],
+  member: DiscordMember,
+  guild: ApiGuild,
+) => {
+  if (guild.ownerId === member.user.id) return true;
+
+  // eslint-disable-next-line no-bitwise
+  const adminRoles = roles.filter(
+    (role) => BigInt(role.permissions) & BigInt(8),
+  );
+
+  return member.roles.some((role) =>
+    adminRoles.some((adminRole) => adminRole.id === role),
   );
 };
